+++ WebFormHTR.Solution/WebFormHTR.Application/Features/Template/DTOs/TemplateDetailDto.cs
@@ -8,8 +8,8 @@ namespace WebFormHTR.Application.Features.Template.DTOs;
 public record TemplateDetailDto(
     Guid Id,
     string Name,
-    float Width,
-    float Height,
+    int Width,
+    int Height,
     TemplateWithoutParentDto? Parent,
     FileDto? File,
     DateTime CreatedAt,
