+++ WebFormHTR.Solution/WebFormHTR.Infrastructure/Services/Scripting/DTOs/PythonResidualDto.cs
@@ -4,7 +4,7 @@ namespace WebFormHTR.Infrastructure.Services.Scripting.DTOs;
 
 public class PythonResidualDto
 {
-    [JsonPropertyName("coords")] public List<float> Coords { get; set; } = [];
+    [JsonPropertyName("coords")] public List<int> Coords { get; set; } = [];
 
     [JsonPropertyName("content")] public string? Content { get; set; }
 }
